# Prevent "database is locked"