--- a/.changeset/limit_the_number_of_parallel_sqlite_connections_to_1_to_prevent_database_is_locked_errors.md
+++ b/.changeset/limit_the_number_of_parallel_sqlite_connections_to_1_to_prevent_database_is_locked_errors.md
@@ -1,0 +1,5 @@
+---
+default: patch
+---
+
+# Limit the number of parallel SQLite connections to 1 to prevent "database is locked" errors.