This should effectively get rid of the `database is locked` error by preventing parallel connections. It's like wrapping a mutex around every database interaction and seems to be the recommended way of doing this.

The benchmarks seem to be mostly unaffected which makes sense considering that they are single-threaded. To see how it performs in practice we'll need to run it on a used host.


Going through https://sqlite.org/faq.html#q6 lead me to https://sqlite.org/threadsafe.html and it seems like it's generally recommended to use the `SQLITE_CONFIG_SINGLETHREAD` threading mode if locking isn't required to speed up performance. However, I haven't been able to figure out how to do that from the Go driver.

Closes https://github.com/SiaFoundation/hostd/issues/643

The before and after results of the existing benchmarks:

Before:
```
goos: darwin
goarch: arm64
pkg: go.sia.tech/hostd/v2/persist/sqlite
cpu: Apple M2 Pro
BenchmarkTrimSectors
BenchmarkTrimSectors-12        	  217003	      5595 ns/op	    217003 sectors	    1843 B/op	      50 allocs/op
BenchmarkV2AppendSectors
BenchmarkV2AppendSectors-12    	  170421	     12802 ns/op	    170421 sectors	    2800 B/op	      43 allocs/op
BenchmarkV2TrimSectors
BenchmarkV2TrimSectors-12      	 1356486	       949.9 ns/op	   1356486 sectors	       0 B/op	       0 allocs/op
BenchmarkVolumeGrow
BenchmarkVolumeGrow-12         	  407421	      3158 ns/op	    407421 sectors	     224 B/op	       8 allocs/op
BenchmarkVolumeShrink
BenchmarkVolumeShrink-12       	  716907	      1891 ns/op	    716907 sectors	       0 B/op	       0 allocs/op
BenchmarkVolumeMigrate
BenchmarkVolumeMigrate-12      	      18	  64587782 ns/op	        18.00 sectors	   11207 B/op	     298 allocs/op
BenchmarkStoreSector
BenchmarkStoreSector-12        	    7747	    160465 ns/op	      7747 sectors	    9243 B/op	     225 allocs/op
BenchmarkReadSector
BenchmarkReadSector-12         	   35526	     31730 ns/op	     35526 sectors	    4782 B/op	      94 allocs/op
PASS
ok  	go.sia.tech/hostd/v2/persist/sqlite	953.837s
```

After:
```
goos: darwin
goarch: arm64
pkg: go.sia.tech/hostd/v2/persist/sqlite
cpu: Apple M2 Pro
BenchmarkTrimSectors
BenchmarkTrimSectors-12        	  235959	      5596 ns/op	    235959 sectors	    1843 B/op	      50 allocs/op
BenchmarkV2AppendSectors
BenchmarkV2AppendSectors-12    	  175136	     12997 ns/op	    175136 sectors	    2800 B/op	      43 allocs/op
BenchmarkV2TrimSectors
BenchmarkV2TrimSectors-12      	 1344559	       934.5 ns/op	   1344559 sectors	       0 B/op	       0 allocs/op
BenchmarkVolumeGrow
BenchmarkVolumeGrow-12         	  410400	      3112 ns/op	    410400 sectors	     224 B/op	       8 allocs/op
BenchmarkVolumeShrink
BenchmarkVolumeShrink-12       	  724045	      1898 ns/op	    724045 sectors	       0 B/op	       0 allocs/op
BenchmarkVolumeMigrate
BenchmarkVolumeMigrate-12      	      18	  66599769 ns/op	        18.00 sectors	   11230 B/op	     298 allocs/op
BenchmarkStoreSector
BenchmarkStoreSector-12        	    7740	    154748 ns/op	      7740 sectors	    9240 B/op	     225 allocs/op
BenchmarkReadSector
BenchmarkReadSector-12         	   35342	     31820 ns/op	     35342 sectors	    4783 B/op	      94 allocs/op
PASS
ok  	go.sia.tech/hostd/v2/persist/sqlite	970.016s
```