@ChrisSchinnerl 2025-04-14
Benchmark results for `BenchmarkStoreSectorParallel`

No limit on connections:
```
goos: darwin
goarch: arm64
pkg: go.sia.tech/hostd/v2/persist/sqlite
cpu: Apple M2 Pro
BenchmarkStoreSectorParallel
BenchmarkStoreSectorParallel/5
BenchmarkStoreSectorParallel/5-12  	    7119	    166676 ns/op	25164.41 MB/s	      7119 sectors	    9259 B/op	     225 allocs/op
BenchmarkStoreSectorParallel/10
BenchmarkStoreSectorParallel/10-12 	    5340	    199449 ns/op	21029.47 MB/s	      5340 sectors	    9267 B/op	     225 allocs/op
BenchmarkStoreSectorParallel/50
BenchmarkStoreSectorParallel/50-12 	    2014	    549619 ns/op	7631.29 MB/s	      2014 sectors	    9374 B/op	     227 allocs/op
PASS
ok  	go.sia.tech/hostd/v2/persist/sqlite	26.978s
```

with single connection:
```
goos: darwin
goarch: arm64
pkg: go.sia.tech/hostd/v2/persist/sqlite
cpu: Apple M2 Pro
BenchmarkStoreSectorParallel
BenchmarkStoreSectorParallel/5
BenchmarkStoreSectorParallel/5-12  	    7743	    203230 ns/op	20638.22 MB/s	      7743 sectors	    9406 B/op	     228 allocs/op
BenchmarkStoreSectorParallel/10
BenchmarkStoreSectorParallel/10-12 	    7278	    167688 ns/op	25012.62 MB/s	      7278 sectors	    9400 B/op	     228 allocs/op
BenchmarkStoreSectorParallel/50
BenchmarkStoreSectorParallel/50-12 	    6670	    168916 ns/op	24830.77 MB/s	      6670 sectors	    9406 B/op	     228 allocs/op
PASS
ok  	go.sia.tech/hostd/v2/persist/sqlite	22.258s
```