@ChrisSchinnerl 2025-04-14
Benchmark results for `BenchmarkReadSectorParallel`

No limit on connections:

```
goos: darwin
goarch: arm64
pkg: go.sia.tech/hostd/v2/persist/sqlite
cpu: Apple M2 Pro
BenchmarkReadSectorParallel
BenchmarkReadSectorParallel/5
BenchmarkReadSectorParallel/5-12  	   33870	     32288 ns/op	129901.93 MB/s	     33870 sectors	    4774 B/op	      94 allocs/op
BenchmarkReadSectorParallel/10
BenchmarkReadSectorParallel/10-12 	   34621	     35217 ns/op	119098.13 MB/s	     34621 sectors	    4784 B/op	      94 allocs/op
BenchmarkReadSectorParallel/50
BenchmarkReadSectorParallel/50-12 	   21943	     49934 ns/op	83997.03 MB/s	     21943 sectors	    4810 B/op	      94 allocs/op
PASS
ok  	go.sia.tech/hostd/v2/persist/sqlite	28.188s
```

With single connection:
```
goos: darwin
goarch: arm64
pkg: go.sia.tech/hostd/v2/persist/sqlite
cpu: Apple M2 Pro
BenchmarkReadSectorParallel
BenchmarkReadSectorParallel/5
BenchmarkReadSectorParallel/5-12  	   36141	     32980 ns/op	127176.83 MB/s	     36141 sectors	    4895 B/op	      97 allocs/op
BenchmarkReadSectorParallel/10
BenchmarkReadSectorParallel/10-12 	   36258	     33593 ns/op	124854.95 MB/s	     36258 sectors	    4890 B/op	      97 allocs/op
BenchmarkReadSectorParallel/50
BenchmarkReadSectorParallel/50-12 	   35840	     33117 ns/op	126652.24 MB/s	     35840 sectors	    4894 B/op	      97 allocs/op
PASS
ok  	go.sia.tech/hostd/v2/persist/sqlite	22.673s
```