@ChrisSchinnerl 2025-04-14
> A less scorched earth design would be to split up read and write transactions so read operations can continue to be parallel since they don't lock the database and writes are forced to be serial.

I can try a `RWMutex` approach or the 2 pool approach where one pool (the one for write txns) is limited to 1 connection and the other one to multiple. My only concern there would be some sort of starvation on the writer side.

Looking at the parallel read benchmark it also seems like even though multiple connections should have the clear advantage, for some reason a single one scales better.